requirements.txt
langchain
langchain-openai
python-dotenv
ipykernel
langchain_community
langchain_core
streamlit
langchain_ollama
langchain-groq
langchain-huggingface
faiss-cpu
sentence-transformers
pypdf
wikipedia
arxiv
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators==0.28.1
youtube_transcript_api
unstructured
pytube
numexpr